Add Platformer Effector. Add more traps

+++ Assets/Script/DoorScript.cs
@@ -4,20 +4,28 @@ using UnityEngine;
 
 public class DoorScript : MonoBehaviour
 {
-    [SerializeField] GameObject playerr;
+    AudioManager audiomanager;
+    [SerializeField] GameObject playerr, levelloader;
     // Start is called before the first frame update
     void Start()
     {
-        
+        audiomanager = GameObject.FindObjectOfType<AudioManager>();
     }
 
     // Update is called once per frame
     void Update()
     {
-        if(Vector2.Distance(playerr.transform.position, transform.position) < 1.5f && Input.GetKeyDown(KeyCode.F)){
+        if(Vector2.Distance(playerr.transform.position, transform.position) < 1.5f && Input.GetKeyDown(KeyCode.F) && playerr.GetComponent<PlayerScript>().KeyCount > 0){
+            StartCoroutine(PlayDoorSound());
             Debug.Log("The door is opened!");
             playerr.GetComponent<PlayerScript>().playerinfohandler.DeleteKey();
             //Load Scene or Moving Camera
+            levelloader.GetComponent<LevelLoader>().LoadNextLevel();
         }
+    }
+    IEnumerator PlayDoorSound(){
+        audiomanager.Play("OpenDoor");
+        yield return new WaitForSeconds(0.5f);
+        audiomanager.Play("CloseDoor");
     }
 }
